Add Stripe checkout and emailing
Now the user is redirected to diff endpoints on etherize.io when they
complete/cancel a fiat payment. I'm not sure how we should save the
users' openlaw form details though if they have cancelled a fiat
payment.

Mailgun emails are working, TODO is add an endpoint for requested
features like the adoption.