Add kaleido ping - may not be sufficient
I added a kaleido ping function that gets an openlaw jwt every 70 hours,
though from Kaleido's documentation it says they pause an instance that
doesn't commit transactions to a block, so we may need to ping a
block-writing function. I also changed the log name to reflect the date
it was created.